Implement scrolling to selected item
While working on this, I found out that the every-layout style of
applying "data-i" to the DOM elements it defines does not work very
well with liveview's DOM diffing. So this also includes a few changes
to the used layouts so that they are controlled more by plain CSS
and less javascript DOM manipulation.